requirements.txt
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.3.4
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==3.3.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
Deprecated==1.2.12
Django==3.2.3
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
entrypoints==0.3
idna==2.10
ipykernel==5.5.5
ipython==7.24.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
jwcrypto==0.9
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
oauthlib==3.1.1
olefile==0.44
packaging==20.9
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.1.0
requests==2.25.1
selenium==3.6.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.16.0
sqlparse==0.4.1
terminado==0.10.0
testpath==0.5.0
tornado==6.1
traitlets==5.0.5
urllib3==1.26.5
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
